Add Motorcycle Features and Transaction Model

+++ app/Http/Controllers/CarController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Car;
+use App\Models\Order;
+use App\Models\Transaction;
 use GuzzleHttp\Psr7\Request;
+use Illuminate\Support\Facades\Date;
 use Intervention\Image\ImageManager as Image;
 use Illuminate\Support\Facades\Storage;
 use League\CommonMark\Extension\CommonMark\Node\Inline\Strong;
@@ -11,8 +14,8 @@ use League\CommonMark\Extension\CommonMark\Node\Inline\Strong;
 class CarController extends Controller
 {
     public function index(){
-        if(request("sort") != NULL){
-            $sort = request("sort");
+        $sort = request("sort");
+        if($sort != NULL){
             if ($sort == "Ear"){
                 $carList = Car::orderBy("created_at", "asc");
             }
@@ -29,32 +32,32 @@ class CarController extends Controller
         else{
             $carList = Car::latest();
         }
-        if (request("model") != NULL) {
-            $carList->whereRaw(sprintf("model regexp '%s'", request("model")));
+        foreach(request()->except("sort", "_token", "search", "startDate", "endDate") as $f => $v){
+            if($v != NULL){
+                $carList->where("$f", $v);
+            }
         }
-        if (request("production_year") != NULL) {
-            $carList->where("production_year", request("production_year"));
-        }
-        if (request("transmission") != NULL) {
-            $carList->where("transmission", request("transmission"));
-        }
-        if (request("type") != NULL) {
-            $carList->where("type", request("type"));
-        }
-        if (request("brand") != NULL) {
-            $carList->where("brand", request("brand"));
-        }
-        if (request("people_capacity") != NULL) {
-            $carList->where("people_capacity", request("people_capacity"));
-        }
-        if (request("engine_type") != NULL) {
-            $carList->where("engine_type", request("engine_type"));
+        $search = request("search");
+        if(request("search") != NULL){
+            $carList->whereAny([
+                "model",
+                "brand"
+            ], "LIKE", "%$search%");
         }
 
+        $startDate = request("startDate");
+        $endDate = request("endDate");
+        if($startDate && $endDate){
+            $orders = Order::where('start_date', '>=', request("startDate"))
+                    ->where('end_date', '<=', request("endDate"))
+                    ->pluck('car_id');
+            $carList->whereNotIn('id', $orders)->get();
+        }
         $carList = $carList->paginate(10);
         return view("cars.index", compact("carList"));
 
     }
+
 
     public function create(){
         if(auth()->user()){
@@ -83,10 +86,10 @@ class CarController extends Controller
                     "image" => "required|image|mimes:jpeg,png,jpg,gif,svg:max:2048",
                     "color" => 'required'
                 ]);
-                $image = request()->file('image');
-                if ($image) {
-                    $imagePath = $image->store('car', 'public');
-                    $validatedFields['image'] = $imagePath;
+                $image = request()->file("image");
+                if($image){
+                    $imagePath = $image->store("car", "public");
+                    $validatedFields["image"] = $imagePath;
                     Car::create($validatedFields);
                 }
                 return redirect(route("cars.index"));
@@ -138,12 +141,18 @@ class CarController extends Controller
     }
     public function destroy(Car $car){
         if(auth()->user()){
-            if(auth()->user()->isAdmin){
+            if(auth()->user()->isAdmin && Car::where("id", $car->id)){
                 Storage::disk("public")->delete($car->image ?? "");
                 $car->delete();
                 return redirect(route("cars.index"));
             }
             return abort(403);
+        }
+        return redirect(route("auth.login"));
+    }
+    public function show(Car $car, Date $startDate, Date $endDate){
+        if(auth()->user() && Car::where("id", $car->id)){
+            return view("cars.show", compact("car", "startDate", "endDate"));
         }
         return redirect(route("auth.login"));
     }
